requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
psycopg==3.2.5
argon2-cffi==23.1.0
